tests/wpt/web-platform-tests/navigation-timing/dom_interactive_media_document.html

<!DOCTYPE html>
<html>
    <head>
        <meta charset="utf-8" />
        <title>Test domInteractive on media document.</title>
        <link rel="author" title="Google" href="http://www.google.com/" />
        <link rel="help" href="https://html.spec.whatwg.org/multipage/browsing-the-web.html#read-media"/>
        <script src="/resources/testharness.js"></script>
        <script src="/resources/testharnessreport.js"></script>
    </head>
    <script>
        const t = async_test("Test domInteractive on media document");
        function frameLoaded() {
            const timing = document.querySelector("iframe").contentWindow.performance.timing;
            assert_greater_than(timing.domInteractive, 0,
                "Expect domInteractive to be positive value.");
            t.done();
        }
    </script>
    <body>
        <h1>Description</h1>
        <p>This tests that a media document has positive-value domInteractive.</p>
        <iframe src="../media/A4.mp4" onload="frameLoaded()"></iframe>
    </body>
</html>
